Arreglos en el servicio de registro de posición y cambio a MapView

+++ common/src/main/java/com/gf/common/entity/RunningUIState.kt
@@ -5,7 +5,7 @@ import com.google.android.gms.maps.model.LatLng
 data class RunningUIState(
     val time : Int,
     val distance : Int,
-    val speedLastKm : Int,
+    val speedLastKm : Float,
     val points : List<List<LatLng>>,
     val status : ActivityStatus
 )
